Extract MCP catalog schema boundary
